requirements.txt
boto==2.49.0
boto3==1.12.20
botocore==1.15.20
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
docutils==0.15.2
environs-9.2.0
fastapi-0.61.2
fuzzywuzzy==0.18.0
gensim==3.8.1
idna==2.9
jmespath==0.9.5
kiwisolver==1.1.0
matplotlib==3.2.0
marshmallow-3.9.1
nltk==3.4.5
numpy==1.18.1
OpenNMT-py==1.1.1
pandas==1.0.1
Pillow==7.0.0
pydantic-1.7.2
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv-0.15.0
python-Levenshtein==0.12.0
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==1.9.0
starlette-0.13.6
torch==1.4.0+cpu
torchvision==0.5.0+cpu
Unidecode==1.1.1
urllib3==1.25.8
